requirements/requirements-packaging.txt
Django==4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-cors-headers==3.11.0
psycopg2-binary==2.9.2
gunicorn==20.1.0
django-rest-passwordreset==1.2.1
djangorestframework-simplejwt==5.0.0
pytz==2021.3
